move static tx to test utils
